Update FreeRTOS header
