+++ library.properties
@@ -1,5 +1,5 @@
 name=FreeRTOS
-version=10.2.0-3
+version=10.2.1-0
 author=Richard Barry <info@freertos.org>
 maintainer=Phillip Stevens <phillip.stevens@gmail.com>
 sentence=Real Time Operating System implemented for AVR (Uno, Leonardo, Mega).
